+++ demo/src/plugins/dashboard/components/top-articles/TableRow.jsx
@@ -5,6 +5,7 @@ import convertReadableTime from '@triniti/cms/utils/convertReadableTime';
 import Message from '@gdbots/pbj/Message';
 import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
 import pbjUrl from '@gdbots/pbjx/pbjUrl';
+import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
 import { Button, Icon, RouterLink } from '@triniti/admin-ui-plugin/components';
 
 const TableRow = ({ idx, article }) => (
@@ -34,20 +35,23 @@ const TableRow = ({ idx, article }) => (
         <Button id={`view-${article.get('_id')}`} color="hover" radius="circle" className="mb-0 mr-1">
           <Icon imgSrc="eye" alt="view" />
         </Button>
+        <UncontrolledTooltip placement="auto" target={`view-${article.get('_id')}`}>View</UncontrolledTooltip>
       </RouterLink>
       <RouterLink to={`${pbjUrl(article, 'cms')}/edit`}>
         <Button id={`edit-${article.get('_id')}`} color="hover" radius="circle" className="mb-0 mr-1">
           <Icon imgSrc="pencil" alt="edit" />
         </Button>
+        <UncontrolledTooltip placement="auto" target={`edit-${article.get('_id')}`}>Edit</UncontrolledTooltip>
       </RouterLink>
       <a
         href={pbjUrl(article, 'canonical')}
         rel="noopener noreferrer"
         target="_blank"
       >
-        <Button color="hover" radius="circle" className="mr-1 mb-0">
+        <Button id={`open-in-new-tab-${article.get('_id')}`} color="hover" radius="circle" className="mr-1 mb-0">
           <Icon imgSrc="external" alt="open" />
         </Button>
+        <UncontrolledTooltip placement="auto" target={`open-in-new-tab-${article.get('_id')}`}>Open in new tab</UncontrolledTooltip>
       </a>
     </td>
   </tr>