+++ src/plugins/ncr/reducers/nodes.js
@@ -1,5 +1,6 @@
 import createReducer from '@triniti/app/createReducer';
 import { actionTypes as pbjxActionTypes } from '@triniti/cms/plugins/pbjx/constants';
+import { actionTypes } from '../constants';
 
 
 export const initialState = {};
@@ -27,6 +28,17 @@ const onPbjxEnvelopeReceived = (prevState, action) => {
   return state;
 };
 
+const onNodesRecieved = (prevState, action) => {
+  const state = { ...prevState };
+  action.nodes.forEach((node) => {
+    const label = node.schema().getCurie().getMessage();
+    state[label] = { ...state[label] };
+    state[label][node.get('_id')] = node.freeze();
+  });
+  return state;
+};
+
 export default createReducer(initialState, {
   [pbjxActionTypes.ENVELOPE_RECEIVED]: onPbjxEnvelopeReceived,
+  [actionTypes.NODES_RECEIVED]: onNodesRecieved,
 });