+++ src/plugins/news/components/article-fields/index.jsx
@@ -10,6 +10,7 @@ import React from 'react';
 import Schema from '@gdbots/pbj/Schema';
 import SponsorPickerField from '@triniti/cms/plugins/boost/components/sponsor-picker-field';
 import TextField from '@triniti/cms/components/text-field';
+import UserPicker from '@triniti/cms/plugins/iam/components/user-picker-field';
 
 const ArticleFields = ({ article, isEditMode, schemas }) => (
   <>
@@ -51,6 +52,13 @@ const ArticleFields = ({ article, isEditMode, schemas }) => (
             picklistId="article-themes"
           />
           )}
+        <FieldArray
+          component={UserPicker}
+          isEditMode={isEditMode}
+          label="Author"
+          name="authorRefs"
+          placeholder="Select an author..."
+        />
       </CardBody>
     </Card>
     <Card>